Update Class to remove timeSlot and use startTime/endTime instead
To more easily interface with the front end, we are now storing a
startTime and endTime separately instead of a startTime and a duration
value that was being used to calculate the endtime. This commit also
includes all fixes to the rest of the backend code to account for this
change.