Implement DELETE studio class route
This commit contains all backend code necessary to implement the
following route: `DELETE /studios/<studio-id>/classes/<class-id> that
deletes a class resource and removes it from the Studio Repository.

TODO: Figure out if we should keep it within the ClassRepository or also
delete it there, currently there isn't a use for the ClassRepository so
will have to see.